@@ -5,7 +5,7 @@
 
     <groupId>eu.europeana.statistics.backend</groupId>
     <artifactId>statistics-dashboard-backend</artifactId>
-    <version>4</version>
+    <version>5</version>
     <packaging>pom</packaging>
 
     <modules>
@@ -17,26 +17,25 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <version.metis>11</version.metis>
-        <version.slf4j>1.7.30</version.slf4j>
-        <version.jul.to.slf4j>1.7.30</version.jul.to.slf4j>
-        <version.log4j>2.17.1</version.log4j>
-        <version.elastic.apm>1.43.0</version.elastic.apm>
-        <version.spring.boot>2.7.18</version.spring.boot>
-        <version.spring>5.3.31</version.spring>
-        <version.swagger.annotations>1.6.2</version.swagger.annotations>
-        <version.jackson>2.12.3</version.jackson>
-        <version.commons.lang3>3.12.0</version.commons.lang3>
         <version.maven.compiler.plugin>3.8.0</version.maven.compiler.plugin>
         <version.maven.release.plugin>2.5.3</version.maven.release.plugin>
         <version.surefire.plugin>3.1.2</version.surefire.plugin>
+        <version.metis>12-SNAPSHOT</version.metis>
+        <version.log4j>2.23.0</version.log4j>
+        <version.slf4j>2.0.12</version.slf4j>
+        <version.elastic.apm>1.48.1</version.elastic.apm>
+        <version.spring.boot>3.2.3</version.spring.boot>
+        <version.spring>6.1.4</version.spring>
         <version.swagger>3.0.0</version.swagger>
+        <version.swagger.annotations>1.6.2</version.swagger.annotations>
+        <version.jackson>2.16.1</version.jackson>
+        <version.commons.lang3>3.12.0</version.commons.lang3>
         <version.servlet.api>4.0.1</version.servlet.api>
-        <version.javax.annotation>1.3.2</version.javax.annotation>
-        <version.jaxb.api>2.3.1</version.jaxb.api>
-        <version.jaxb.runtime>2.3.2</version.jaxb.runtime>
-        <version.junit>5.8.1</version.junit>
-        <version.mockito>5.6.0</version.mockito>
+        <version.jakarta.annotation>2.1.1</version.jakarta.annotation>
+        <version.jakarta.xml.bind-api>4.0.1</version.jakarta.xml.bind-api>
+        <version.jaxb.runtime>4.0.4</version.jaxb.runtime>
+        <version.junit>5.10.2</version.junit>
+        <version.mockito>5.8.0</version.mockito>
         <version.jetbrains>24.0.1</version.jetbrains>
     </properties>
 
@@ -104,13 +103,20 @@
                     <artifactId>maven-compiler-plugin</artifactId>
                     <version>${version.maven.compiler.plugin}</version>
                     <configuration>
-                        <release>17</release>
+                        <release>21</release>
                     </configuration>
                 </plugin>
                 <plugin>
                     <groupId>org.apache.maven.plugins</groupId>
                     <artifactId>maven-surefire-plugin</artifactId>
                     <version>${version.surefire.plugin}</version>
+                    <configuration>
+                        <systemPropertyVariables>
+                            <!--Temporary argument to bypass exception on inline mockito mocks(coming from transitive bytebuddy dependency).-->
+                            <!--Remove this when mockito has been updated to work with java 21(At time of writing 5.8.0 does not work).-->
+                            <net.bytebuddy.experimental>true</net.bytebuddy.experimental>
+                        </systemPropertyVariables>
+                    </configuration>
                 </plugin>
             </plugins>
         </pluginManagement>
