@@ -1,7 +1,6 @@
 package eu.europeana.statistics.dashboard.rest.config;
 
 import com.mongodb.client.MongoClient;
-import eu.europeana.corelib.web.socks.SocksProxy;
 import eu.europeana.metis.mongo.connection.MongoClientProvider;
 import eu.europeana.metis.mongo.connection.MongoProperties;
 import eu.europeana.metis.utils.CustomTruststoreAppender;
@@ -9,11 +8,10 @@ import eu.europeana.metis.utils.CustomTruststoreAppender.TrustStoreConfiguration
 import eu.europeana.metis.utils.apm.ElasticAPMConfiguration;
 import eu.europeana.statistics.dashboard.service.StatisticsService;
 import eu.europeana.statistics.dashboard.service.persistence.MongoSDDao;
-import javax.annotation.PreDestroy;
+import jakarta.annotation.PreDestroy;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
@@ -21,11 +19,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 /**
  * The web application making available the repository functionality. This provides all the
@@ -33,12 +26,11 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
  * required setup.
  */
 @Configuration
-@EnableWebMvc
 @Import({ElasticAPMConfiguration.class})
 @ComponentScan(basePackages = {"eu.europeana.statistics.dashboard.rest.controller",
         "eu.europeana.statistics.dashboard.rest.exception"})
 @EnableScheduling
-public class ApplicationConfiguration implements WebMvcConfigurer{
+public class ApplicationConfiguration {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(ApplicationConfiguration.class);
 
@@ -67,17 +59,11 @@ public class ApplicationConfiguration implements WebMvcConfigurer{
   static MongoClient initializeApplication(ConfigurationPropertiesHolder propertiesHolder)
           throws TrustStoreConfigurationException {
 
-    // Set the SOCKS proxy
-    if (propertiesHolder.isSocksProxyEnabled()) {
-      new SocksProxy(propertiesHolder.getSocksProxyHost(), propertiesHolder.getSocksProxyPort(),
-              propertiesHolder.getSocksProxyUsername(), propertiesHolder.getSocksProxyPassword()).init();
-    }
-
     // Set the truststore.
     LOGGER.info("Append default truststore with custom truststore");
     if (StringUtils.isNotEmpty(propertiesHolder.getTruststorePath())
             && StringUtils.isNotEmpty(propertiesHolder.getTruststorePassword())) {
-      CustomTruststoreAppender.appendCustomTrustoreToDefault(propertiesHolder.getTruststorePath(),
+      CustomTruststoreAppender.appendCustomTruststoreToDefault(propertiesHolder.getTruststorePath(),
               propertiesHolder.getTruststorePassword());
     }
 
@@ -94,11 +80,6 @@ public class ApplicationConfiguration implements WebMvcConfigurer{
     return new MongoClientProvider<>(mongoProperties).createMongoClient();
   }
 
-  @Override
-  public void addCorsMappings(CorsRegistry registry) {
-    registry.addMapping("/**").allowedMethods("GET", "HEAD", "POST", "PUT", "DELETE", "OPTIONS")
-        .allowedOrigins(properties.getAllowedCorsHosts());
-  }
 
   @Bean
   public MongoSDDao getMongoSDDao(){
@@ -118,19 +99,6 @@ public class ApplicationConfiguration implements WebMvcConfigurer{
   public void refreshRightsUrlCategoryMapping() {
     statisticsService.refreshRightsUrlsCategoryMapping();
   }
-
-  @Override
-  public void addResourceHandlers(ResourceHandlerRegistry registry) {
-    registry.addResourceHandler("/swagger-ui/**")
-            .addResourceLocations("classpath:/META-INF/resources/webjars/springfox-swagger-ui/")
-            .resourceChain(false);
-  }
-
-  @Override
-  public void addViewControllers(ViewControllerRegistry registry) {
-    registry.addRedirectViewController("/", "/swagger-ui/index.html");
-  }
-
   /**
    * Closes any connections previous acquired.
    */
